Landslide commit that is the new multiGSEA
This is a huge breaking change to the code base and was worked on for > 2 months
on a local git repository on my machine. There is no hope to reconcile the
differences between versions 0.1.x and this current version (0.2.x).

The previous/historic working version of this package along with the version
of `rmd.plugins` that worked with it which were being used to generate reports
up until 2015-01-13 can be found in SVN here:

  * http://resscm.gene.com/bioinfo/projects/R/tags/multiGSEA_pre-refactor/
  * http://resscm.gene.com/bioinfo/projects/R/tags/rmd.plugins_pre-refactor/



git-svn-id: http://resscm.gene.com/bioinfo/projects/R/trunk/multiGSEA@156611 ea8427e4-251f-0410-985a-fb3ea31d2ee2