Adds support for DGEList input and romer method.
Note that all logFC statistics are performed via a limma (+voom) pipeline.

If the input expression object is a `DGEList`, the GSEA methods will be ones that
are defined as *.DGElist (ie. camera.DGEList), but the DGEList will be voomed
and processed via limma for the log fold changes statistics.

git-svn-id: http://resscm.gene.com/bioinfo/projects/R/trunk/multiGSEA@170249 ea8427e4-251f-0410-985a-fb3ea31d2ee2