Adds `with.kegg=FALSE` parameter to `getMSigGeneSetDb` function.
Since the KEGG pathways are under a restrictive license, we're doing everybody a
favor by not including them by default when a user asks for the MSigDB
collection. If `with.kegg=TRUE` the KEGG pathways will be returned.

Will write a getKeggGeneSetDb() function that returns KEGG genesets.
The user will be able to specify to use the latest c2 genesets or the
ones provided in the `KEGG.db`, which represents the state of the database
before it converted to a commercial license.

git-svn-id: http://resscm.gene.com/bioinfo/projects/R/trunk/multiGSEA@206979 ea8427e4-251f-0410-985a-fb3ea31d2ee2