Insures packages are installed for methods invoked via multiGSEA()
In order not to inflate the package count required to install multiGSEA, many GSEA
methods that multiGSEA come from packages listed in the `Suggests` field. The `method`
parameter is know checked prior to running any analyses, and the `multiGSEA` call will
error-out cleanly if a required package is missing.

Prior to this commit, the methods that depend on a suggested package would simply error
out, but the multiGSEA call would return a `MultiGSEAResult` object which "looked" ok
as long as you didn't ask for results from the method that failed. This commit now
removes any trace of methods that fail during execution so that the `MultiGSEAResult`
object is kosher.

This commit fixes #5.

Thanks to @tomsing1 for raising issue and helping to debug.