Adds a workaround to support 0-sd features for eigenWeightedMean and gsdScore
I tried a few ways to rescue a signature matrix with 0-sd features, but this
suggested implenetation was largely as robust as other methods, and much simpler.

I think I'd still suggest end users to call single-sample-scoring methods through
the `scoreSingleSamples` function, which pre-emptively removes 0-sd features out
of the gate, but there are times users might want to call the individual methods
directly and not have them blow up (ie. when called through the facileverse,
for instance)

Thanks to @darsoo for the suggested fix.
Closes #20